requirements.txt
torch
torchvision
torchaudio
torchrl
tensordict 
tqdm
numpy
pandas
wandb
gymnasium
hydra-core
ipykernel